Add fuse mixallgather op,  Add sclice tensor op, Fix gcc82 error (#39)

+++ paddle/fluid/framework/data_feed.h
@@ -840,7 +840,7 @@ class SlotObjAllocator {
       delete tmp;
       --capacity_;
     }
-    CHECK_EQ(capacity_, 0);
+    CHECK_EQ(capacity_, static_cast<size_t>(0));
   }
   T* acquire(void) {
     T* x = NULL;
